Revert "chore: module based file structure for /video group"
