+++ apps/web/app/future/video/meeting-ended/[uid]/page.tsx
@@ -1,9 +1,9 @@
+import Page from "@pages/video/meeting-ended/[uid]";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import { _generateMetadata } from "app/_utils";
 import { WithLayout } from "app/layoutHOC";
 
-import MeetingEnded from "~/videos/views/videos-meeting-ended-single-view";
-import { getServerSideProps } from "~/videos/views/videos-meeting-ended-single-view.getServerSideProps";
+import { getServerSideProps } from "@lib/video/meeting-ended/[uid]/getServerSideProps";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
@@ -13,4 +13,4 @@ export const generateMetadata = async () =>
 
 const getData = withAppDirSsr(getServerSideProps);
 
-export default WithLayout({ getData, Page: MeetingEnded, getLayout: null })<"P">;
+export default WithLayout({ getData, Page, getLayout: null })<"P">;
